Separate deserialization information from model file
